backend/requirements.txt
fastapi[standard]==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.2
pydantic-settings==2.5.2

langgraph==0.2.60
langchain-anthropic==0.3.12
langchain-core==0.3.53

pystac-client==0.8.5
planetary-computer==1.0.0
rasterio==1.4.3
numpy==2.1.3
shapely==2.0.6
geojson==3.1.0
Pillow==11.0.0
httpx==0.28.1

python-dotenv==1.0.1
websockets==13.1
